# docs/swagger.yaml
basePath: /
definitions:
  app.conceptSimilaritySearchHeader:
    properties:
      agent_id:
        type: string
    type: object
  app.conceptSimilaritySearchPayload:
    properties:
      embedded_text:
        type: string
      embedding_vector:
        items:
          type: number
        type: array
      search_metrics:
        type: string
      top_k:
        type: integer
    type: object
  app.conceptSimilaritySearchRequest:
    properties:
      header:
        $ref: '#/definitions/app.conceptSimilaritySearchHeader'
      payload:
        $ref: '#/definitions/app.conceptSimilaritySearchPayload'
      request_id:
        type: string
    type: object
  app.conceptSimilaritySearchResponse:
    properties:
      error:
        type: string
      header:
        $ref: '#/definitions/app.conceptSimilaritySearchResponseHeader'
      response_id:
        type: string
      results:
        items:
          $ref: '#/definitions/app.conceptSimilaritySearchResult'
        type: array
      status:
        type: string
    type: object
  app.conceptSimilaritySearchResponseHeader:
    properties:
      agent_id:
        type: string
      mas_id:
        type: string
      workspace_id:
        type: string
    type: object
  app.conceptSimilaritySearchResult:
    properties:
      concept_id:
        type: string
      concept_name:
        type: string
      embedding_vector:
        items:
          type: number
        type: array
      score:
        type: number
    type: object
info:
  contact: {}
  description: IoC Cognition Fabric Node service — shared memory routing and memory
    operations proxy.
  title: CFN Service API
  version: "1.0"
paths:
  /api/internal/workspaces/{workspaceId}/multi-agentic-systems/{masId}/concepts/similarity-search:
    post:
      consumes:
      - application/json
      description: Finds concept nodes nearest to a query embedding vector using HNSW
        index.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Include raw embedding vectors in results (debug only)
        in: query
        name: include_embeddings
        type: boolean
      - description: Similarity search request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app.conceptSimilaritySearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.conceptSimilaritySearchResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Concept similarity search
      tags:
      - Graph Store
  /api/internal/workspaces/{workspaceId}/multi-agentic-systems/{masId}/graph/concepts/by_ids:
    post:
      consumes:
      - application/json
      description: Returns concept details for each of the provided concept IDs.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Concept IDs request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch concepts by IDs
      tags:
      - Graph Store
  /api/internal/workspaces/{workspaceId}/multi-agentic-systems/{masId}/graph/neighbors/{conceptId}:
    get:
      description: Returns the neighboring concepts of a given concept in the knowledge
        graph.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Concept ID
        in: path
        name: conceptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get neighbors by concept ID
      tags:
      - Graph Store
  /api/internal/workspaces/{workspaceId}/multi-agentic-systems/{masId}/graph/paths:
    post:
      consumes:
      - application/json
      description: Returns ordered paths through the knowledge graph between a source
        and target concept.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Graph paths request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch paths between two concepts
      tags:
      - Graph Store
...
swagger: "2.0"